City/requirements.txt
anyio==4.13.0
asgiref==3.11.1
CacheControl==0.14.4
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==48.0.0
dj-database-url==3.1.2
Django==6.0.5
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
firebase_admin==7.4.0
google-api-core==2.30.3
google-auth==2.52.0
google-cloud-core==2.6.0
google-cloud-firestore==2.27.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.9.0
googleapis-common-protos==1.75.0
grpcio==1.80.0
grpcio-status==1.80.0
gunicorn==26.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.15
msgpack==1.1.2
packaging==26.2
proto-plus==1.28.0
protobuf==6.33.6
psycopg2==2.9.12
psycopg2-binary==2.9.12
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.34.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.7.0
whitenoise==6.12.0
